+++ lib/zenlish/wclasses/adverb.rb
@@ -3,7 +3,7 @@ require_relative 'word_class'
 module Zenlish
   module WClasses
     # TODO: document
-    class Adverb < WordClass
+    class Adverb < WordClass   
     end # class
   end # module
 end # module