+++ lib/zenlish/wclasses/comparative_particle.rb
@@ -4,7 +4,7 @@ module Zenlish
   module WClasses
     # TODO: document 
     # Example: 'as' in 'same ... as'
-    class ComparativeParticle < WordClass
+    class ComparativeParticle < WordClass   
     end # class
   end # module
 end # module