+++ lib/zenlish/wclasses/determiner.rb
@@ -3,7 +3,7 @@ require_relative 'word_class'
 module Zenlish
   module WClasses
     # TODO: document
-    class Determiner < WordClass
+    class Determiner < WordClass   
     end # class
   end # module
 end # module