+++ lib/zenlish/wclasses/fronting_quantifier.rb
@@ -3,7 +3,7 @@ require_relative 'quantifier'
 module Zenlish
   module WClasses
     # Quantifiers that appear in front of determiners: all, both
-    class  FrontingQuantifier < Quantifier
+    class  FrontingQuantifier < Quantifier    
     end # class
   end # module
 end # module