+++ lib/zenlish/wclasses/preposition.rb
@@ -3,12 +3,7 @@ require_relative 'word_class'
 module Zenlish
   module WClasses
     # TODO: document
-    class Preposition < WordClass
-    
-      # Prepositions do not inflect, so they're invariable.
-      def invariable?
-        true
-      end    
+    class Preposition < WordClass   
     end # class
   end # module
 end # module