+++ lib/zenlish/wclasses/pronoun.rb
@@ -3,7 +3,7 @@ require_relative 'word_class'
 module Zenlish
   module WClasses
     # TODO: document
-    class Pronoun < WordClass
+    class Pronoun < WordClass    
     end # class
   end # module
 end # module