+++ lib/zenlish/wclasses/quantifier.rb
@@ -4,7 +4,7 @@ module Zenlish
   module WClasses
     # Quantifiers are used to indicate the number or quantity of the noun under consideration.
     # enough, few, little, many, more, most, much, none, several, some
-    class  Quantifier < Determiner
+    class  Quantifier < Determiner   
     end # class
   end # module
 end # module