+++ spec/zenlish/inflect/inflection_table_builder_spec.rb
@@ -174,33 +174,32 @@ module Zenlish
           expect(table.inflect(verb_4, [nil,    nil,     :past_participle, nil])).to eq('caused')          
         end
         
-        # MockIrregularVerb = Struct.new(:base_form, :PERSON, :NUMBER, :TIME)
-        # it 'should build default paradigm for irregular verbs' do
-          # table = subject.build('Irregular_form') do
-            # feature_heading 'PERSON'
-            # feature_heading 'NUMBER'
-            # feature_heading 'TIME'
-            # method_heading  'base_form'
-            # #     PERSON             NUMBER             TIME                         base_form
-            # rule([not_equal(:third), dont_care,         equals(:present),            dont_care], col('base_form'))
-            # rule([equals(:third),    equals(:singular), equals(:present),            matches(/[^aeiouy]y$/)], sub(col('base_form'), /y$/, 'ies'))
-            # rule([equals(:third),    equals(:singular), equals(:present),            matches(/(?:[osxz]|ch|sh)$/)], concat(col('base_form'), 'es'))             
-            # rule([equals(:third),    equals(:singular), equals(:present),            dont_care], concat(col('base_form'), 's'))         
-            # rule([dont_care,         dont_care, equals(:progressive),                matches(/e$/)], sub(col('base_form'), /e$/, 'ing'))            
-            # rule([dont_care,         dont_care, equals(:progressive),                dont_care], concat(col('base_form'), 'ing'))
-            # rule([dont_care,         dont_care, in?(:past_simple, :past_participle), matches(/e$/)], concat(col('base_form'), 'd'))             
-            # rule([dont_care,         dont_care, in?(:past_simple, :past_participle), matches(/[^aeiouy]y$/)], sub(col('base_form'), /y$/, 'ied'))            
-            # rule([dont_care,         dont_care, in?(:past_simple, :past_participle), dont_care], concat(col('base_form'), 'ed'))
-          # end
+        MockIrregularVerb = Struct.new(:base_form, :PERSON, :NUMBER, :TIME, :past_simple, :past_participle)
+        it 'should build default paradigm for irregular verbs' do
+          table = subject.build('Irregular_inflection') do
+            feature_heading 'PERSON'
+            feature_heading 'NUMBER'
+            feature_heading 'TIME'
+            method_heading  'base_form'
+            #     PERSON             NUMBER             TIME                      base_form
+            rule([not_equal(:third), dont_care,         equals(:present),         dont_care], col('base_form'))
+            rule([equals(:third),    equals(:singular), equals(:present),         matches(/[^aeiouy]y$/)], sub(col('base_form'), /y$/, 'ies'))
+            rule([equals(:third),    equals(:singular), equals(:present),         matches(/(?:[osxz]|ch|sh)$/)], concat(col('base_form'), 'es'))             
+            rule([equals(:third),    equals(:singular), equals(:present),         dont_care], concat(col('base_form'), 's'))         
+            rule([dont_care,         dont_care,         equals(:progressive),     matches(/e$/)], sub(col('base_form'), /e$/, 'ing'))            
+            rule([dont_care,         dont_care,         equals(:progressive),     dont_care], concat(col('base_form'), 'ing'))
+            rule([dont_care,         dont_care,         equals(:past_simple),     dont_care], func('past_simple'))             
+            rule([dont_care,         dont_care,         equals(:past_participle), dont_care], func('past_participle'))                       
+          end
 
-          # verb_1 = MockIrregularVerb.new('choose', :first, :singular, :present)
-          # expect(table.inflect(verb_1, [nil,    nil,     nil, nil]             )).to eq('choose')
-          # expect(table.inflect(verb_1, [:third, nil,     nil, nil]             )).to eq('chooses')
-          # expect(table.inflect(verb_1, [nil,    :plural, nil, nil]             )).to eq('choose')
-          # expect(table.inflect(verb_1, [nil,    nil,     :progressive, nil]    )).to eq('choosing')
-          # expect(table.inflect(verb_1, [nil,    nil,     :past_simple, nil]    )).to eq('chose')
-          # expect(table.inflect(verb_1, [nil,    nil,     :past_participle, nil])).to eq('chosen')
-        # end
+          verb_1 = MockIrregularVerb.new('choose', :first, :singular, :present, 'chose', 'chosen')
+          expect(table.inflect(verb_1, [nil,    nil,     nil, nil]             )).to eq('choose')
+          expect(table.inflect(verb_1, [:third, nil,     nil, nil]             )).to eq('chooses')
+          expect(table.inflect(verb_1, [nil,    :plural, nil, nil]             )).to eq('choose')
+          expect(table.inflect(verb_1, [nil,    nil,     :progressive, nil]    )).to eq('choosing')
+          expect(table.inflect(verb_1, [nil,    nil,     :past_simple, nil]    )).to eq('chose')
+          expect(table.inflect(verb_1, [nil,    nil,     :past_participle, nil])).to eq('chosen')
+        end
       end # context
     end # describe
   end # module